+++ src/Hero.java
@@ -1,6 +1,6 @@
 abstract class Hero extends DungeonCharacter{
 
-    public Hero(String theName, int theHitPoints, int theMinDam, int theMaxDam, int theAttackSpeed, int theHitChance) {
+    public Hero(String theName, int theHitPoints, int theMinDam, int theMaxDam, int theAttackSpeed, double theHitChance) {
         super(theName, theHitPoints, theMinDam, theMaxDam, theAttackSpeed, theHitChance);
     }
